Seperate out validator integration tests
